@@ -162,7 +162,7 @@ class _NotificationsMoreScreenState extends State<NotificationsMoreScreen> with 
             Container(
               padding: const EdgeInsets.all(8),
               decoration: BoxDecoration(
-                color: ChessEarnTheme.getColor('brand-accent')?.withOpacity(0.1),
+                color: ChessEarnTheme.getColor('brand-accent').withOpacity(0.1),
                 borderRadius: BorderRadius.circular(8),
               ),
               child: Icon(
